+++ apiphany-core/src/main/java/org/apiphany/RequestParameters.java
@@ -1,7 +1,5 @@
 package org.apiphany;
 
-import java.lang.annotation.Annotation;
-import java.lang.reflect.AnnotatedElement;
 import java.net.URLDecoder;
 import java.nio.charset.Charset;
 import java.util.ArrayList;
@@ -9,12 +7,10 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.function.Function;
 
 import org.apiphany.http.URIEncoder;
 import org.apiphany.lang.Require;
 import org.apiphany.lang.Strings;
-import org.apiphany.lang.collections.JavaArrays;
 import org.apiphany.lang.collections.Lists;
 import org.apiphany.lang.collections.Maps;
 import org.apiphany.openapi.MultiValueStrategy;
@@ -22,7 +18,7 @@ import org.apiphany.openapi.QueryParam;
 import org.morphix.convert.MapConversions;
 import org.morphix.convert.function.SimpleConverter;
 import org.morphix.convert.strategy.ConversionStrategy;
-import org.morphix.lang.Nullables;
+import org.morphix.lang.JavaArrays;
 import org.morphix.lang.function.PutFunction;
 import org.morphix.reflection.Constructors;
 import org.morphix.reflection.ExtendedField;
@@ -245,47 +241,26 @@ public class RequestParameters {
 	 * @return a map representation of the object's fields
 	 */
 	protected static Map<String, List<String>> fromObject(final Object queryParams, final SimpleConverter<String, String> nameConverter) {
-		List<ExtendedField> fields = ConversionStrategy.findFields(queryParams,
-				field -> null != field.getFieldValue());
+		List<ExtendedField> extendedFields = ConversionStrategy.findFields(queryParams,
+				extendedField -> null != extendedField.getFieldValue());
 
-		Map<String, List<String>> paramMap = new LinkedHashMap<>(fields.size());
-		if (Lists.isEmpty(fields)) {
+		Map<String, List<String>> paramMap = new LinkedHashMap<>(extendedFields.size());
+		if (Lists.isEmpty(extendedFields)) {
 			return paramMap;
 		}
-		for (ExtendedField extendedField : fields) {
+		for (ExtendedField extendedField : extendedFields) {
 			List<String> paramValues = RequestParameter.toValues(extendedField.getFieldValue());
 			if (Lists.isEmpty(paramValues)) {
 				continue;
 			}
 			String paramName = nameConverter.convert(extendedField.getName());
 
-			QueryParam annotation = getAnnotation(extendedField, QueryParam.class);
+			QueryParam annotation = extendedField.getAnnotation(QueryParam.class);
 			MultiValueStrategy strategy = MultiValueStrategy.from(annotation);
 			ParameterFunction parameter = strategy.apply(paramName, paramValues);
 			parameter.putInto(paramMap);
 		}
 		return paramMap;
-	}
-
-	/**
-	 * Retrieves the annotation from the given {@link ExtendedField}, if present. The method checks both the field and its
-	 * getter method for the annotation, giving precedence to the field annotation if both are present.
-	 * <p>
-	 * TODO: move this method to ExtendedField and make it more generic to retrieve any annotation.
-	 *
-	 * @param <T> the type of the annotation to retrieve
-	 *
-	 * @param extendedField the {@link ExtendedField} to check for the annotation
-	 * @param annotationClass the class of the annotation to retrieve
-	 * @return the {@link QueryParam} annotation if present, or {@code null} if not found
-	 */
-	private static <T extends Annotation> T getAnnotation(final ExtendedField extendedField, final Class<T> annotationClass) {
-		Function<AnnotatedElement, T> getAnnotation = annotated -> annotated.getAnnotation(annotationClass);
-		T annotation = Nullables.apply(extendedField.getField(), getAnnotation);
-		if (null != annotation) {
-			return annotation;
-		}
-		return Nullables.apply(extendedField.getGetterMethod(), getAnnotation);
 	}
 
 	/**