+++ apiphany-core/src/main/java/org/apiphany/http/HttpHeader.java
@@ -4,6 +4,7 @@ import java.util.Map;
 import java.util.Objects;
 
 import org.morphix.lang.Enums;
+import org.morphix.lang.function.ToStringFunction;
 
 /**
  * RFC 7231 HTTP headers.
@@ -488,7 +489,7 @@ public enum HttpHeader {
 	/**
 	 * The name map for easy from string implementation.
 	 */
-	private static final Map<String, HttpHeader> NAME_MAP = Enums.buildNameMap(values(), header -> header.toString().toLowerCase());
+	private static final Map<String, HttpHeader> NAME_MAP = Enums.buildNameMap(values(), ToStringFunction.toLowerCase());
 
 	/**
 	 * The {@link String} value.
@@ -541,5 +542,4 @@ public enum HttpHeader {
 	public String toString() {
 		return value();
 	}
-
 }